Potential fix for code scanning alert no. 34

+++ drivers/mmc/fsl_esdhc.c
@@ -468,7 +468,7 @@ static int esdhc_send_cmd_common(struct fsl_esdhc_priv *priv, struct mmc *mmc,
 			 * speculative pre-fetching etc.
 			 */
 			dma_unmap_single(priv->dma_addr,
-					 data->blocks * data->blocksize,
+					 (size_t)data->blocks * data->blocksize,
 					 mmc_get_dma_dir(data));
 		}
 	}
